requirements.txt
cffi==1.16.0
cryptography==41.0.5
et-xmlfile==1.1.0
mysql-connector-python==8.2.0
openpyxl==3.1.2
protobuf==4.21.12
pycparser==2.21
pyOpenSSL==23.3.0
secure-smtplib==0.1.1
